Change ID AUtomation/requirements.txt
playwright==1.49.0
python-dotenv==1.0.1
schedule==1.2.2
